# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin-show-detail-order/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Detail a order by ID",
                "operationId": "8654127e6d7d298fa3c33cafd71a57b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to detail",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/admin-update-status-order/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Update a status order by ID",
                "operationId": "43770b848db31c3412ffa4037be003f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_status"
                                ],
                                "properties": {
                                    "order_status": {
                                        "type": "string",
                                        "example": "ordered"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "order not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}